versao sem aldrabar

+++ printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tborges- <tborges-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 17:12:44 by tborges-          #+#    #+#             */
-/*   Updated: 2024/05/26 22:38:21 by tborges-         ###   ########.fr       */
+/*   Updated: 2024/05/26 23:17:07 by tborges-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	print_str(char *str);
 int	print_ptr(unsigned long long ptr);
 int	print_nbr(int nbr);
 int	print_unsigned(unsigned int num);
-int	print_hex(unsigned long long num, const char specifier);
+int	print_hex(unsigned int num, const char specifier);
 int	print_percent(void);
 
 #endif
